feat: dependencies, types & folder structure
