+++ src/.gitignore
@@ -28,7 +28,7 @@ yarn-error.log*
 # local env files
 .env*.local
 .env*
-
+.env
 
 # vercel
 .vercel
