+++ src/backend/db/prisma/schema.prisma
@@ -21,10 +21,9 @@ model UserProfile {
   email                    String
   password                 String
   birthDate                DateTime
-  myaddress                String
-  secondAddress            String?
+  primaryAddress           String
+  secondaryAddress         String?
   createdAt                DateTime             @default(dbgenerated("now()"))
-  isActive                 Boolean              @default(true)
   createdGroups            Group[]              @relation("CreatedBy") // list of created groups
   groupMember              GroupMember[]        @relation("groupMember") // list of groups that user belongs to
   wishlists                Wishlist[] // wishlists created by this user
