+++ src/backend/index.ts
@@ -1,10 +1,11 @@
-import express, { Request, Response } from 'express';
+// THIS FILE IS TO INITIALIZE THE SERVER WITH EXPRESS
+import express, { Request, Response } from "express";
 
 const app = express();
 const port = 3000;
 
-app.get('/', (req: Request, res: Response) => {
-  res.send('Hello World!');
+app.get("/", (req: Request, res: Response) => {
+  res.send("Hello World!");
 });
 
 app.listen(port, () => {
