+++ src/backend/routes/users/models.ts
@@ -1,0 +1,1 @@
+// THESE ARE THE MODELS FOR USERS
